+++ tests/testthat/test-metric-silhouette.R
@@ -1,4 +1,4 @@
-test_that("multiplication works", {
+test_that("silhouette works", {
   kmeans_spec <- k_means(num_clusters = 1) %>%
     set_engine("stats")
 
@@ -15,4 +15,22 @@ test_that("multiplication works", {
     sil_width = rep(NA_real_, 32)
   )
   expect_identical(res, exp_res)
+})
+
+test_that("silhouette() errors for cluster spec", {
+  spec <- tidyclust::k_means(num_clusters = 4)
+
+  expect_snapshot(
+    error = TRUE,
+    silhouette(spec)
+  )
+})
+
+test_that("silhouette_avg() errors for cluster spec", {
+  spec <- tidyclust::k_means(num_clusters = 4)
+
+  expect_snapshot(
+    error = TRUE,
+    silhouette_avg(spec)
+  )
 })